// src/api/schema.d.ts
export interface components {
    schemas: {
// ...
        BoardTask: {
            id: string;
            title: string;
            number: number | null;
            description: string | null;
            status: string;
            /** @description One of: no-priority, low, medium, high, urgent. */
            priority: string;
            /** Format: date-time */
            startDate: string | null;
            /** Format: date-time */
            dueDate: string | null;
            position: number | null;
            /** Format: date-time */
            createdAt: string;
            userId: string | null;
            assigneeName: string | null;
            assigneeId: string | null;
            assigneeImage: string | null;
            projectId: string;
            labels: components["schemas"]["TaskLabel"][];
            externalLinks: components["schemas"]["TaskExternalLink"][];
        };
// ...
        BoardColumn: {
            /** @description The column slug, same as `slug`. */
            id: string;
            slug: string;
            name: string;
            icon: string | null;
            isFinal: boolean;
            tasks: components["schemas"]["BoardTask"][];
        };
// ...
        BoardResponse: {
            data: components["schemas"]["Board"];
            pagination: components["schemas"]["BoardPagination"];
        };
        Board: {
            id: string;
            name: string;
            slug: string;
            icon: string | null;
            description: string | null;
            isPublic: boolean | null;
            workspaceId: string;
            columns: components["schemas"]["BoardColumn"][];
            archivedTasks: components["schemas"]["BoardTask"][];
            plannedTasks: components["schemas"]["BoardTask"][];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}